Rewrite Element#method_missing to allow BFS
In order to make breadth first search actually work, a number of
other recently made changes had to be made first, and then I 
still needed to rewrite the searching code to support searching
through a queue that is changing in each iteration.

Also rewrite the documentation associated with #method_missing
to be more up to date and more comprehensive.

Signed-off-by: Mark Rada <mrada@marketcircle.com>