Modify Element#method_missing to account for non-existant classes
In some cases, the class being searched for will not exist when
the initial class constant lookup is made. So we have to modify
the algorithm to update the cached class constant every time that
we wrap new children (because wrapping new children might generate
the class we are looking for).

Signed-off-by: Mark Rada <mrada@marketcircle.com>