Screw gemsets
They are only making things more difficult. 

It is good to separate the development state of a project from the production version with a meta-project, but when it comes to deploying the sub-project to the other sub-projects, having
gemsets for multiple projects is just a pain in the ass.

Bundler is enough.

Signed-off-by: Mark Rada <mrada@marketcircle.com>