Get rid of jeweler dependency
This requires me to rewrite tasks that I use, but otherwise removes several gems that would
need to be installed for development.

This also means I have to keep the gemspec up to date myself, which should not be difficult,
given that it is almost identical to what jeweler added to the rakefile.

Signed-off-by: Mark Rada <mrada@marketcircle.com>