Expand notifications to work in synchronous cases
The reason behind this is conceptual. Technically, the only time that a notification is
needed is when working an asynchronous event, but there are also other cases where it
feels natural to use a notification even though they are not needed.

I think it is important to support these cases in order to keep test implementers from
having to think about whether an action is executed asynchronously or not since that
seems more like an implementation detail.

Signed-off-by: Mark Rada <mrada@marketcircle.com>