Introduce Accessibility::Tree
A new way of iterating through a UI hierarchy. The advantages over the current
implementation are that it offers better performance in some edge cases right now,
and overall better performance in conjunction with some tweaks in the long run.

This design also is better for encapsulation, and by mixing in the Enumerable module
we open more options for enumerating over the hierarchy later in different and more
powerful ways.