Handle cases where filter attributes do not exist on some candidate objects
Even though they are the correct type they may not have the attribute that we are
trying to filter, so we can instantly fail in that case.

One example is filtering on placeholder_value, which in the Safari window is used by
the address bar and search box, but not always used by the text fields on the rendered web page.