Rewrite Language#set to support new features
The set method now has an implicit key of 'value' which is what was being set most of
the time.

The second change is that we will try to set focus to an element before trying to
change the attribute value; the reason for this is explained in the documentation 
note.

Also update the documentation.