Use #get_attribute instead of #attribute in searches
Turns out it is faster because constant lookups are faster, which makes up for the
extra check that would have been performed.

Constant lookups are faster now only because we are memoizing the attribute lookup
results and because the lookup table is in the local namespace and not the outermost
namespace (which is where all Cocoa constants are).