Remote invalid todo item
It is invalid because we should let the element decide where the mouse should be 
instead of figuring it out in #move_mouse_to. This allows different implementations
of #to_point to cause the mouse to move to a different place, for example, we may
want #to_point to return something bottom right corner for a GrowArea, or over top the
scroll arrows for a ScrollBar.