Implement nested searching
Using a syntax like:

	app.window( button: { title: 'Add' } )

What will happen is that the a search from app will be started to find a window that
has a button that is titled 'Add'. Searches can be nested arbitrarily and along with
other search filters.

I have yet to add tests, but I have confirmed that it works in the required case.