Refactor Qualifier#meets_criteria? to be less coupled
So that it will be easier to support the existing set of comparison inferences. That
is, by doing this, we can move the logic for comparing the table header title and the
title UI element value into the respective classes.

A good change as it means the logic is less complicated, which in turns means there
is less work for the computer to do and it will run a little faster. The downside is
that the logic is more distributed instead of contained; however, I think that in
this case it is safe to make the logic have a more global scope.