Refactor StringParser to be a mixin where needed
Otherwise Accessibility::Core is doing more than it should. As a mixin, we can also
avoid having to allocate a new instance for each string that needs to be parsed.

I still need to fix a now broken test by moving it over to the string parser. Which
means I should also finish the hotkeys stuff at the same time. :)