Put vendored ActiveSupport::Inflector in Accessibility namespace
Gems like 'mail' try to detect whether it should load ActiveSupport or its own
equivalent based on whether ActiveSupport is already defined in the runtime. Since
AXElements only has vendored the parts of ActiveSupport that it needs it would not
be safe to try and expose it for other dependencies to use.