Make mouse movements more accurate by being less accurate
To compensate for the inaccuracies of MacRuby floats, Mouse has been
changed from always performing a set number of animation steps to
performing as many animation steps as needed to get within a distance
of 2 pixels from the desired goal.

This sounds silly, but it seems like the only reasonable way to handle
the fact that screen co-ordinates are double precision and MacRuby handles
double precision poorly (in terms of precision). The other consequence is
that the duration of an animation may be slightly longer.

I was thinking that an extra animation step should be added to set the
position exactly to where it is supposed to go, but I figured that the
inaccuracy is closer to simulating a human. :)
set exactly