Add an arbitrary buffer time after mouse events in DSL
Often times it seems we need to add a buffer after clicking just to wait for UI to update, in most
cases it is a fraction of a second. Forcing at least 0.2 seconds of waiting should likely reduce
boiler plate without causing a noticeable slow down in the few cases where it is not needed.