Make Element#set:to: automagically work with Ranges
Range objects which do not use negative indices are handled in Core
as of fb94f7fe07. If you use negative indices you would have to add
the extra step of using Range#relative_to yourself to convert the
range to one without negative indices.

At the Element layer of abstraction, we can handle the conversion
process automatically so that it Just Works™.