Remove unneeded sqrt() calculation
The #close_enough method does not need to find
the true distance between points since we want
to be within 1 unit of distance. The trick is to
realize that square roots will always approach 1 no
matter which side of the number line they are on.

Mouse animations will be closer in timings now than
previously.