Overhaul Core
Handle AttributeUnsupported as an error case as we must check. This
is due to the inconsistent implementations of accessibility stuff in
Cocoa.

The error handling methods are now implemented as lambdas and everything
is stored in the AXERROR constant. This is a bit slower, but looks
cleaner.

Instead of treating Accessibility::Core as a mixin, we have
Accessibility::Element that wraps an AXUIElementRef.

Returned arrays are wrapped concurrently to improve performance. This
particular optimization will have to make its way up to the next
abstraction layer.