Replace Factory with an Accessibility::Element extension
Factory was only used for handling elements and arrays of
elements. This has been replaced by a more object oriented
interface that does not need to perform any type checking.

It also has the benefit of automatically taking advantage of
the naively concurrent implementation of NSArray#to_ruby that
is already in use.

This also required a change in AX::Element to make the appropriate
call of #to_ruby. This will add an extra overhead of a function
call to any attribute lookup, however, the method call overhead
is relatively small.