Add an extra note to the README about verbose tests
Instead of running tests in alphabetical order, which
can still be a pain, run them in the normal random
order, but print out which test is currently running.

In order to do this, you will need to run tests by
calling `ruby test_thing.rb` instead of using a rake
task. At least for now.