Fix some MacRuby specific hacks that AX::Element depended on
Namely calling String methods on Symbol objects. In
MacRuby, Symbol is a subclass of NSString and due
to the unified runtime and laziness, Symbol will
respond to any method that NSString will respond to.

Since NSString needs to mimic the behaviour of a
Ruby String, MacRuby added many methods to NSString.

The combination of these two details allowed me
to call String#chomp on a Symbol in MacRuby in
order to avoid a call to #to_s and make the code
look a bit prettier. 

So the optimal solution is to add a freedom patch
on MRI that adds Symbol#chomp